Separated out profiler code
