+++ morpho5/vm/profile.h
@@ -1,0 +1,15 @@
+/** @file profile.h
+ *  @author T J Atherton
+ *
+ *  @brief Profiler
+ */
+
+#ifndef profile_h
+#define profile_h
+
+#include "compile.h"
+#include "vm.h"
+#include "profile.h"
+#include "morpho.h"
+
+#endif /* profile_h */
